requirements.txt
yt-dlp
faster-whisper
openai>=1.0.0
psycopg2-binary
pgvector
click
python-dotenv
